fix: refactor private _infra imports
