+++ qualtran/surface_code/flasq/span_counting_test.py
@@ -16,7 +16,7 @@
 import pytest
 import sympy
 
-from qualtran._infra.composite_bloq import BloqBuilder
+from qualtran import BloqBuilder
 from qualtran.bloqs.basic_gates import CNOT, Hadamard, Swap, Toffoli
 from qualtran.bloqs.mcmt import And
 from qualtran.resource_counting._costing import get_cost_value
